requirements.txt
flask==3.0.0
line-bot-sdk==3.7.0
python-dotenv==1.0.0
gunicorn==21.2.0
dropbox==11.36.2
requests==2.31.0 